correct margins, move eventCard and eventCardContainer into components folder
