+++ Cohort2/team6/client/src/components/EventCard/EventCard.test.jsx
