+++ Cohort2/team6/client/src/pages/Home/Home.tsx
@@ -1,6 +1,6 @@
 import { Box, Heading } from "@chakra-ui/react";
 import Carousel from "../Home/components/Carousel/Carousel";
-import EventCardContainer from "../Home/components/EventCardContainer/EventCardContainer";
+import EventCardContainer from "../../components/EventCardContainer/EventCardContainer";
 import { EventTypes } from "../../types/Event.types";
 export const Home = () => {
   // dummy data until redux store is setup
